app
good

+++ src/js/modules/links.js
@@ -11,11 +11,16 @@ function app_Enter_Register(){
     const password_confirm = document.getElementById("register-Password-C").value;
     const grade = document.getElementById("register-Grade").value;
     const section = document.getElementById("register-Section").value;
-    localStorage.setItem("name", nme)
-    localStorage.setItem("lastName", lastName)
-    localStorage.setItem("password", password)
-    localStorage.setItem("password_C", password_confirm) 
-    localStorage.setItem("grade", grade)
-    localStorage.setItem("section", section)
-    window.location.href="/src/views/apk.html" 
+    if (password == password_confirm) {
+        localStorage.setItem("name", nme)
+        localStorage.setItem("lastName", lastName)
+        localStorage.setItem("password", password)
+        localStorage.setItem("password_C", password_confirm) 
+        localStorage.setItem("grade", grade)
+        localStorage.setItem("section", section)
+        window.location.href="/src/views/apk.html"
+    }else{
+     let x = document.getElementById('test_password')
+        x.innerHTML= `<p>La contraseña no es igual a la primera </p>`
+    }
 }
